requirements.txt
python-telegram-bot==20.7
google-genai==1.0.0
pdf2image==1.17.0
Pillow==10.4.0
pymongo==4.10.1
weasyprint==62.3
reportlab==4.0.7
